+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+## [1.52.0] - 2025-07-05
+
 ### Added
 
 - Add a warning message to `luau-lsp analyze` when `--platform=roblox` is set but no definitions files are provided